src/test/java/app/daos/MovieDAOTest.java
package app.daos;

import app.dtos.MovieDTO;
import app.entities.Cast;
import app.entities.Genre;
import app.entities.Movie;
import config.HibernateConfig;
import jakarta.persistence.EntityManager;
import jakarta.persistence.EntityManagerFactory;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import java.lang.reflect.Array;
import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;

import static org.junit.jupiter.api.Assertions.*;

class MovieDAOTest {

    private static MovieDAO movieDAO;

    private static EntityManagerFactory emfTest;
    private static EntityManager em;

    @BeforeAll
    static void beforeAll() {


        emfTest = HibernateConfig.getEntityManagerFactoryConfig(true);
        movieDAO = MovieDAO.getInstance(emfTest);

    }


    @BeforeEach
    void beforeEach() {
        em = emfTest.createEntityManager();

    }

    @AfterEach
    void tearDown() {
        em.close();
    }

    @Test
    void searchForReleaseDate() {


        List<String> originCountry = new ArrayList<>();
        originCountry.add("UK");
        originCountry.add("USA");
        originCountry.add("DK");
        originCountry.add("UK");


        Genre action = new Genre("Action");
        Genre horror = new Genre("Horror");

        List<Genre> genres = new ArrayList<>();
        genres.add(action);
        genres.add(horror);

        List<Cast> casts = new ArrayList<>();

        Movie movie1 = new Movie(1, originCountry, "Transformer", "2020-04-06", genres);
        Movie movie2 = new Movie(2, originCountry, "Transformer2", "2024-09-09", genres);
        Movie movie3 = new Movie(3, originCountry, "Transformer3", "2026-01-01", genres);

        List<Movie> movielist = new ArrayList<>();
        movielist.add(movie1);
        movielist.add(movie2);

        casts.add(new Cast(1, "bob", "Actor", "Egon", movielist));

        try (EntityManager em = emfTest.createEntityManager()) {
            em.getTransaction().begin();
            em.persist(action);
            em.persist(horror);
            em.persist(movie1);
            em.persist(movie2);
            em.persist(movie3);
            em.getTransaction().commit();

            List<Movie> findMoviesExpectedTrue1 = movieDAO.searchForReleaseDate("2020-04-06");
            List<Movie> findMoviesExpectedTrue2 = movieDAO.searchForReleaseDate("2024-09-09");
            List<Movie> findMoviesExpectedTrue3 = movieDAO.searchForReleaseDate("2026-01-01");


            Movie findMovie1 = em.find(Movie.class, 1);
            Movie findMovie2 = em.find(Movie.class, 2);
            Movie findMovie3 = em.find(Movie.class, 3);


            assertEquals(findMovie1.getId(), findMoviesExpectedTrue1.get(0).getId());
            assertEquals(findMovie1.getTitle(), findMoviesExpectedTrue1.get(0).getTitle());
            assertEquals(findMovie1.getReleaseDate(), findMoviesExpectedTrue1.get(0).getReleaseDate());
            assertEquals(findMovie2.getReleaseDate(), findMoviesExpectedTrue2.get(0).getReleaseDate());
            assertEquals(findMovie3.getReleaseDate(), findMoviesExpectedTrue3.get(0).getReleaseDate());

                // ------------------- List objects not working ? -----------------------
//            assertEquals(findMovie1.getGenres(), findMoviesExpectedTrue1.get(0).getGenres()); // DOES NOT WORK
//            assertEquals(findMovie1.getCasts(), findMoviesExpectedTrue1.get(0).getCasts()); // DOES NOT WORK
//            assertEquals(findMovie1.getOriginCountry(), findMoviesExpectedTrue1.get(0).getOriginCountry()); // DOES NOT WORK


        }


    }

    @Test
    void seeAllMovies() {
    }

    @Test
    void persistDTOasEntity() {
    }

    @Test
    void persistCast() {
    }

    @Test
    void persistList() {
    }

}